feat: Use H2 Database when maintain Question
- 이제 Question을 관리할 때 H2 DB를 이용합니다.
- id를 관리할 때, strategy가 GenerationType.AUTO이면 Question과 User가
자동생성되는 id 값을 공유합니다. 그래서, GenerationType.IDENTITY로
변경하였습니다.
- 일단, Question 내에서 writer를 User로 두는 것을 다시 단순한 String으로
변경하였습니다.
