+++ src/main/java/com/codessquad/qna/domain/QuestionRepository.java
@@ -1,0 +1,6 @@
+package com.codessquad.qna.domain;
+
+import org.springframework.data.repository.CrudRepository;
+
+public interface QuestionRepository extends CrudRepository<Question, Long> {
+}
