--- src/main/java/com/codessquad/qna/question/Question.java
@@ -1,62 +1,0 @@
-package com.codessquad.qna.question;
-
-import com.codessquad.qna.domain.User;
-
-import java.time.LocalDateTime;
-
-public class Question {
-    private int id;
-    private User writer;
-    private String title;
-    private String contents;
-    private LocalDateTime time;
-    private int point;
-
-    public int getId() {
-        return id;
-    }
-
-    public void setId(int id) {
-        this.id = id;
-    }
-
-    public User getWriter() {
-        return writer;
-    }
-
-    public void setWriter(User writer) {
-        this.writer = writer;
-    }
-
-    public String getTitle() {
-        return title;
-    }
-
-    public void setTitle(String title) {
-        this.title = title;
-    }
-
-    public String getContents() {
-        return contents;
-    }
-
-    public void setContents(String contents) {
-        this.contents = contents;
-    }
-
-    public LocalDateTime getTime() {
-        return time;
-    }
-
-    public void setTime(LocalDateTime time) {
-        this.time = time;
-    }
-
-    public int getPoint() {
-        return point;
-    }
-
-    public void setPoint(int point) {
-        this.point = point;
-    }
-}
