--- src/main/java/com/codessquad/qna/question/QuestionRequest.java
@@ -1,32 +1,0 @@
-package com.codessquad.qna.question;
-
-public class QuestionRequest {
-    private String writer;
-    private String title;
-    private String contents;
-
-    public void setWriter(String writer) {
-        this.writer = writer;
-    }
-
-    public void setTitle(String title) {
-        this.title = title;
-    }
-
-    public void setContents(String contents) {
-        this.contents = contents;
-    }
-
-    public String getWriter() {
-        return writer;
-    }
-
-    public String getTitle() {
-        return title;
-    }
-
-    public String getContents() {
-        return contents;
-    }
-
-}
