+++ src/main/resources/templates/index.html
@@ -13,7 +13,7 @@
               <div class="auth-info">
                 <i class="icon-add-comment"></i>
                 <span class="time">{{formatDateTime time}}</span>
-                <a href="/users/{{writer.userId}}" class="author">{{writer.name}}</a>
+                <a href="/users/{{writer}}" class="author">{{writer}}</a>
               </div>
               <div class="reply" title="댓글">
                 <i class="icon-reply"></i>
