+++ src/main/resources/templates/qna/show.html
@@ -12,7 +12,7 @@
                           <img src="https://graph.facebook.com/v2.3/100000059371774/picture" class="article-author-thumb" alt="">
                       </div>
                       <div class="article-header-text">
-                          <a href="/users/{{question.writer.userId}}" class="article-author-name">{{question.writer.name}}</a>
+                          <a href="/users/{{question.writer}}" class="article-author-name">{{question.writer}}</a>
                           <a href="/questions/{{question.id}}" class="article-header-time" title="퍼머링크">
                               {{formatDateTime question.time}}
                               <i class="icon-link"></i>
